// resume_builder_front/src/types/user.ts
import type { IDateValue, ISocialLink } from './common';

// ----------------------------------------------------------------------
export type UserStatus = 'Active' | 'Inactive';
// export type IUserTableFilters = {
//   name: string;
//   role: string[];
//   status: string;
// };

export type IUserTableFilters = {
  name: string;
  role: string[];
  status: 'all' | UserStatus;
};

export type IUserProfileCover = {
  name: string;
  role: string;
  coverUrl: string;
  avatarUrl: string;
};

export type IUserProfile = {
  id: string;
  role: string;
  quote: string;
  email: string;
  school: string;
  country: string;
  company: string;
  totalFollowers: number;
  totalFollowing: number;
  socialLinks: ISocialLink;
};

export type IUserProfileFollower = {
  id: string;
  name: string;
  country: string;
  avatarUrl: string;
};

export type IUserProfileGallery = {
  id: string;
  title: string;
  imageUrl: string;
  postedAt: IDateValue;
};

export type IUserProfileFriend = {
  id: string;
  name: string;
  role: string;
  avatarUrl: string;
};

export type IUserProfilePost = {
  id: string;
  media: string;
  message: string;
  createdAt: IDateValue;
  personLikes: { name: string; avatarUrl: string }[];
  comments: {
    id: string;
    message: string;
    createdAt: IDateValue;
    author: { id: string; name: string; avatarUrl: string };
  }[];
};

export type IUserAPI = {
  id: number;
  firstname: string;
  lastname: string;
  fathername: string;
  username: string;
  email: string;
  phone: string;
  position: string;
  gender: string;
  dob: string;
  createDate: string;
  profileImage: string;
  isActive: boolean;
  nid: string;
};

export type IUserCard = {
  id: number;
  name: string;
  role: string;
  coverUrl: string;
  avatarUrl?: string;
  totalPosts: number;
  totalFollowers: number;
  totalFollowing: number;
};

export type IUserItem = {
  id: string;
  name: string;
  city: string;
  role: string;
  email: string;
  state: string;
  status: string;
  address: string;
  country: string;
  zipCode: string;
  company: string;
  avatarUrl: string;
  phoneNumber: string;
  isVerified: boolean;
};
export type IUserItemListView = {
  id: string;
  firstname: string;
  lastname: string;
  fathername: string;
  username: string;
  email: string;
  phone: string;
  position: string;
  password?: string;
  gender: string;
  dob: string;
  profileImage: string;
  isActive: UserStatus;
  nid: string;
};

export type IUserAccountBillingHistory = {
  id: string;
  price: number;
  invoiceNumber: string;
  createdAt: IDateValue;
};